nexus/requirements.txt
pandas
sentence-transformers
faiss-cpu
openai
neo4j
streamlit
python-dotenv
matplotlib 
fastapi
uvicorn 